move postgres related types to a different module
